MAINT move filereader and filewriter to dicomio
