+++ pydicom/dicomio/filereader.py
