+++ pydicom/tests/test_RLE_transfer_syntax.py
@@ -3,7 +3,7 @@ import sys
 import re
 import pytest
 import pydicom
-from pydicom.filereader import read_file
+from pydicom.dicomio import read_file
 from pydicom.data import get_testdata_files
 
 pillow_missing_message = ("pillow is not available "
