+++ pydicom/tests/test_pillow_pixel_data.py
@@ -3,7 +3,7 @@ import os
 import sys
 import pytest
 import pydicom
-from pydicom.filereader import read_file
+from pydicom.dicomio import read_file
 from pydicom.data import get_testdata_files
 from pydicom.tag import Tag
 pillow_missing_message = ("pillow is not available "
